customer can create a new card

+++ DopravniPodnik/DopravniPodnik/Data/Models/TypyPredplatneho.cs
@@ -12,7 +12,7 @@ public partial class TypyPredplatneho
 
     public string Jmeno { get; set; } = null!;
 
-    public decimal Cena { get; set; }
+    public int Cena { get; set; }
 
     public virtual ICollection<Predplatne> Predplatnes { get; set; } = new List<Predplatne>();
 }
